+++ SimpleTester/BenchLockTest.cs
@@ -6,14 +6,14 @@ using BTDB.Locks;
 
 namespace SimpleTester;
 
-// | Method                     | Mean      | Error     | StdDev    |
-// |--------------------------- |----------:|----------:|----------:|
-// | SeqLock_Read               |  1.239 ns | 0.0044 ns | 0.0039 ns |
-// | ReaderWriterLockSlim_Read  | 12.508 ns | 0.0304 ns | 0.0284 ns |
-// | MonitorLock_Read           |  8.579 ns | 0.0291 ns | 0.0272 ns |
-// | SeqLock_Write              |  4.148 ns | 0.0252 ns | 0.0236 ns |
-// | ReaderWriterLockSlim_Write | 13.532 ns | 0.0323 ns | 0.0302 ns |
-// | MonitorLock_Write          |  8.145 ns | 0.0272 ns | 0.0254 ns |
+// | Method                     | Mean       | Error     | StdDev    |
+// |--------------------------- |-----------:|----------:|----------:|
+// | SeqLock_Read               |  0.8039 ns | 0.0120 ns | 0.0100 ns |
+// | ReaderWriterLockSlim_Read  | 12.5249 ns | 0.0526 ns | 0.0466 ns |
+// | MonitorLock_Read           |  8.6376 ns | 0.0472 ns | 0.0441 ns |
+// | SeqLock_Write              |  3.7167 ns | 0.0157 ns | 0.0139 ns |
+// | ReaderWriterLockSlim_Write | 13.5983 ns | 0.0503 ns | 0.0420 ns |
+// | MonitorLock_Write          |  8.1922 ns | 0.0323 ns | 0.0287 ns |
 
 [InProcess]
 public class BenchLockTest